requirements.txt
boto3==1.34.154
botocore==1.34.154
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
jmespath==1.0.1
moviepy==1.0.3
numpy==2.0.1
pillow==10.4.0
proglog==0.1.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
tqdm==4.66.5
urllib3==2.2.2
